ImageScanner/requirements.txt
# **Coffee&Code Document Intelligence Requirements.txt** #
# Core packages
opencv-python==4.11.0.86        # Image processing
pytesseract==0.3.13           # Tesseract OCR wrapper
numpy==2.2.4                 # Array handling

# Interface options (choose one or both)
streamlit==1.44.0              # For web app version
jupyter==1.0.0                  # For notebook version
matplotlib==3.7.1               # For Jupyter visualizations

# Additional dependencies
pillow==11.1.0                  # Image handling for Tesseract
python-dotenv==1.0.0            # For environment variables (optional)